# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Identity Platform API",
    "version": "1.0.0",
    "description": "Wire-level API contract for the Identity Platform. The semantic guarantees for these endpoints live in docs/contracts."
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment"
    },
    {
      "url": "https://auth.example.com",
      "description": "Production example"
    },
    {
      "url": "http://localhost:5300",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered with a token pair",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/InvalidClient" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login succeeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Rotate a refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Revoke a refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token revoked when it exists for the client",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/auth/oauth/{provider}": {
      "get": {
        "tags": ["OAuth"],
        "summary": "Start an OAuth flow",
        "parameters": [
          { "$ref": "#/components/parameters/OAuthProvider" },
          { "name": "client_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "redirect_uri", "in": "query", "required": true, "schema": { "type": "string", "format": "uri" } },
          { "name": "code_challenge", "in": "query", "required": false, "schema": { "type": "string", "minLength": 43, "maxLength": 128 } },
          { "name": "code_challenge_method", "in": "query", "required": false, "schema": { "type": "string", "enum": ["S256"] } }
        ],
        "responses": {
          "302": { "description": "Redirect to the provider authorization URL" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/auth/oauth/{provider}/callback": {
      "get": {
        "tags": ["OAuth"],
        "summary": "Handle provider callback",
        "parameters": [
          { "$ref": "#/components/parameters/OAuthProvider" },
          { "name": "code", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "state", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "error", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "302": { "description": "Redirect to the registered application callback" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/oauth/token": {
      "post": {
        "tags": ["OAuth"],
        "summary": "Exchange an OAuth authorization code",
        "description": "Failed client, redirect URI or PKCE checks do not consume the authorization code. A successful exchange consumes it atomically.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OAuthTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use the form: ApiKey sk_..."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    },
    "parameters": {
...
      "OAuthProvider": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "enum": ["google", "github"] }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InvalidClient": {
        "description": "Invalid client credentials",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "headers": {
          "X-RateLimit-Limit": { "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
          "X-RateLimit-Reset": { "schema": { "type": "integer" } }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "array", "items": {} }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "ClientCredentialFields": {
        "type": "object",
        "required": ["clientId"],
        "properties": {
          "clientId": { "type": "string", "example": "cl_example" },
          "clientSecret": { "type": "string", "example": "cs_example" }
        }
      },
      "RegisterRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/ClientCredentialFields" },
          {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
              "email": { "type": "string", "format": "email", "maxLength": 320 },
              "password": { "type": "string", "minLength": 8, "maxLength": 128 }
            }
          }
        ]
      },
      "LoginRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/ClientCredentialFields" },
          {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
              "email": { "type": "string", "format": "email" },
              "password": { "type": "string" }
            }
          }
        ]
      },
      "RefreshRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/ClientCredentialFields" },
          {
            "type": "object",
            "required": ["refreshToken"],
            "properties": {
              "refreshToken": { "type": "string" }
            }
          }
        ]
      },
...
      "OAuthTokenRequest": {
        "type": "object",
        "required": ["code", "clientId", "redirectUri"],
        "properties": {
          "code": { "type": "string" },
          "clientId": { "type": "string" },
          "clientSecret": { "type": "string" },
          "redirectUri": { "type": "string", "format": "uri" },
          "codeVerifier": { "type": "string", "minLength": 43, "maxLength": 128 }
        }
      },
      "AuthUser": {
        "type": "object",
        "required": ["id", "email"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "required": ["user", "accessToken", "refreshToken", "expiresIn"],
        "properties": {
          "user": { "$ref": "#/components/schemas/AuthUser" },
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "integer", "example": 900 }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "required": ["accessToken", "refreshToken", "expiresIn"],
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "integer", "example": 900 }
        }
      },
...
    }
  }
}